+++ Data/DeathlogData~Init.lua
@@ -1,4 +1,4 @@
-local VERSION = 1778371200
+local VERSION = 1779667200
 
 if DeathlogData and (DeathlogData.VERSION or 0) >= VERSION then return end