+++ Libs/DeathNotificationLib/Data/DeathNotificationLibData~Init.lua
@@ -1,4 +1,4 @@
-local VERSION = 1778371200
+local VERSION = 1779667200
 
 if DeathNotificationLibData and (DeathNotificationLibData.VERSION or 0) >= VERSION then return end